backend/requirements.txt
Flask
python-dotenv
requests
google-generativeai
google-cloud-vision
flask-cors
supabase